requirements.txt
python-twitter=1.3.1
xmpppy==0.5.0rc1
